feat(decomp): recover Player movement and collision
Recover normal, jump, and climb movement; rotation and walk-number display; the Player collision engine; swap/reset paths; and target-backed Player work fields. Correct the Coin motion-number ABI.

Object evidence: player.o is 150/165 paired, 117 strict-exact, and 130 data-value-exact with target/source text 0xBEB0/0x80CC. coin.o is 24 paired and 6 exact with text 0x4550/0x103C. Serialized object builds and git diff --check pass.

Player and Coin remain NonMatching; no full-DOL gate or promotion is claimed.

+++ include/game/board/coin.h
@@ -13,8 +13,8 @@ typedef struct MbCoinObj_s {
 
 s16 mbCoinCreate(void);
 s16 mbCoinCreate2(void);
-s16 mbCoinObjCreate(int modelNo, s8 motNo);
-s16 mbCoinObjCreate2(int modelNo, s8 motNo);
+s16 mbCoinObjCreate(int modelNo, int motNo);
+s16 mbCoinObjCreate2(int modelNo, int motNo);
 MBCOINOBJ *mbCoinObjGet(s16 objId);
 void mbCoinObjNumDec(s16 objId);
 void mbCoinObjKill(s16 objId);
@@ -32,7 +32,7 @@ float mbCoinObjAlphaGet(s16 objId);
 void mbCoinObjDispSet(s16 objId, BOOL dispF);
 BOOL mbCoinObjDispGet(s16 objId);
 void mbCoinObjLayerSet(s16 objId, u16 layer);
-void mbCoinObjMotSet(s16 objId, s8 motNo);
+void mbCoinObjMotSet(s16 objId, int motNo);
 void mbCoinDispKill(s16 no);
 BOOL mbCoinDispKillCheck(s16 no);
